ed25519-donna: Remove duplicate static code
