+++ ed25519-donna/ed25519-donna.h
@@ -19,17 +19,6 @@
 typedef unsigned char hash_512bits[64];
 
 /*
-	Timing safe memory compare
-*/
-static int
-ed25519_verify(const unsigned char *x, const unsigned char *y, size_t len) {
-	size_t differentbits = 0;
-	while (len--)
-		differentbits |= (*x++ ^ *y++);
-	return (int) (1 & ((differentbits - 1) >> 8));
-}
-
-/*
  * Arithmetic on the twisted Edwards curve -x^2 + y^2 = 1 + dx^2y^2
  * with d = -(121665/121666) = 37095705934669439343138083508754565189542113879843219016388785533085940283555
  * Base point: (15112221349535400772501151409588531511454012693041857206046113283949847762202,46316835694926478169428394003475163141307993866256225615783033603165251855960);
